+++ config/hash-generator.php
@@ -6,7 +6,7 @@ return [
     /**
      * @link https://www.php.net/manual/ru/function.openssl-encrypt.php
      */
-    'cipher_algo' => 'rc4', // @link https://www.php.net/manual/ru/function.openssl-get-cipher-methods.php
+    'cipher_algo' => 'aes-128-cfb', // @link https://www.php.net/manual/ru/function.openssl-get-cipher-methods.php
     'passphrase'  => 'pass_phrase', // @link https://www.php.net/manual/ru/function
 //.openssl-random-pseudo-bytes.php
     'options'     => 0, // @link https://www.php.net/manual/ru/function.openssl-encrypt.php